+++ src/index.ts
@@ -1,0 +1,5 @@
+import registerTwitch from './clients/twitch';
+import registerDiscord from './clients/discord';
+
+registerTwitch();
+registerDiscord();
